+++ lib/src/rsp_util.dart
@@ -131,15 +131,16 @@ class Rsp {
    * Turn it off if you have multiple dart files in the same Web page.
    */
   static String script(HttpConnect connect, String src, [bool bootstrap=true]) {
+    final String prefix = connect.server.uriVersionPrefix;
     int i = src.lastIndexOf('.dart'), j = i + 5;
     if (i < 0 || (j < src.length && src[j] != '?'))
-      return '<script src="$src"></script>\n';
+      return '<script src="$prefix$src"></script>\n';
 
     if (disableDartScript || !connect.browser.dart)
-      return '<script src="${src.substring(0,j)}.js${src.substring(j)}"></script>\n';
+      return '<script src="$prefix${src.substring(0,j)}.js${src.substring(j)}"></script>\n';
 
-    final s = '<script type="application/dart" src="$src"></script>\n';
+    final s = '<script type="application/dart" src="$prefix$src"></script>\n';
     return bootstrap ?
-      s + '<script src="/packages/browser/dart.js"></script>\n': s;
+      s + '<script src="$prefix/packages/browser/dart.js"></script>\n': s;
   }
 }